requirements.txt
flask
python-dotenv
flask-wtf
wtforms
flask-sqlalchemy
flask-migrate
flask-login
email-validator
flask-mail
pyjwt
flask-bootstrap
flask-moment
flask-babel
